Fix LEDs. Still need to understand PIO shift out register, RGB vs BRG vs GRB bit ordering

+++ include/led_module.h
@@ -57,7 +57,7 @@ void showLEDs(); // Push buffer to LEDs
 void testLEDs(); // Run LED test sequence
 
 // Helper functions for common colors
-RGBColor RGB(uint8_t r, uint8_t g, uint8_t b);
+RGBColor RGB(uint32_t r, uint32_t g, uint32_t b);
 RGBColor colorRed();
 RGBColor colorGreen();
 RGBColor colorBlue();
